% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADIPOCYTE_SIZE_STATS}
\alias{ADIPOCYTE_SIZE_STATS}
\title{Analysis of adipocyte size differences}
\format{
A \code{data.frame} with 40 rows and 11 variables.
}
\usage{
ADIPOCYTE_SIZE_STATS
}
\description{
Analysis of adipocyte size differences
}
\keyword{datasets}
