% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ANALYTES}
\alias{ANALYTES}
\title{Plasma clinical analytes}
\format{
An object of class \code{data.frame} with 146 rows and 20 columns.

\describe{
\item{\code{plate}}{integer; 96-well plate identifier (1–4).}
\item{\code{runseq}}{integer; run order (1–146).}
\item{\code{platepos}}{integer; position in each 96-well plate.}
\item{\code{pid}}{integer; randomly generated 8-digit identifier used in
linkage to phenotypic data. All samples from the same animal have the same
PID.}
\item{\code{bid}}{integer; unique 5 digit identifier of all samples
collected for an acute test/sample collection period. All samples collected
during that period will have the same BID.}
\item{\code{viallabel}}{integer; character, unique 11 digit numeric
identifier of a sample vial.}
\item{\code{omics_subset}}{logical; whether the sample was selected for
-omics analysis.}
\item{\code{sex}}{factor; the sex of the rat with levels "Female" and
"Male".}
\item{\code{timepoint}}{factor; exercise training group. Either "SED"
(sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").}
\item{\code{corticosterone}}{numeric; corticosterone (ng/mL).}
\item{\code{glucagon}}{numeric; glucagon (pmol).}
\item{\code{glucose}}{numeric; glucose (mg/dL).}
\item{\code{glycerol}}{numeric; glycerol (mg/dL).}
\item{\code{insulin}}{numeric; insulin (pg/mL).}
\item{\code{insulin_iu}}{numeric; insulin (\eqn{\mu}IU/mL) using the
conversion 1 pg/mL = 0.023 \eqn{\mu}IU/mL provided in the package insert.}
\item{\code{insulin_pm}}{numeric; insulin (pM) calculated by dividing
insulin (pg/mL) by its molecular weight (5.804 kDa).}
\item{\code{lactate}}{numeric; lactate (mmol/L).}
\item{\code{leptin}}{numeric; leptin (pg/mL).}
\item{\code{nefa}}{numeric; non-esterified fatty acids (NEFA) (mmol/L).}
\item{\code{total_ketones}}{numeric; total ketone bodies (\eqn{\mu}mol/L).}
}
}
\usage{
ANALYTES
}
\description{
A set of 9 common clinical analytes measured in plasma.
}
\details{
\tabular{lll}{
\bold{Analyte} \tab \bold{Catalogue No.} \tab \bold{Company} \cr
Corticosterone \tab 55-CORMS-E01 \tab Alpco (Salem, NH) \cr
Glucose \tab B24985 \tab Beckman Coulter (Brea, CA) \cr
Glycerol \tab 445850 \tab Beckman Coulter \cr
Lactate \tab A95550 \tab Beckman Coulter \cr
Total Ketones \tab 415-73301, 411-73401 \tab Fujifilm Wako (Osaka, Japan) \cr
NEFA \tab 995-34791, 999-34691, 993-35191, 991-34891 \tab Fujifilm Wako \cr
Glucagon \tab K1535YK \tab Meso Scale  Discovery (Rockville, MD) \cr
Insulin, Leptin \tab K15158C (multiplex assay) \tab Meso Scale Discovery \cr
}
}
\seealso{
\code{\link{ANALYTES_EMM}}, \code{\link{ANALYTES_STATS}}
}
\keyword{datasets}
