% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ANALYTES_EMM}
\alias{ANALYTES_EMM}
\title{Plasma clinical analytes estimated marginal means}
\format{
An object of class \code{list} of length 2.
}
\usage{
ANALYTES_EMM
}
\description{
A nested list of \code{emmGrid} objects. The first level
  indicates whether the EMMs were for timewise or male vs. female (MvF)
  comparisons. The second level is a list of \code{emmGrid} objects, one for
  each clinical analyte: Glucagon, Glucose, Glycerol, Leptin, and NEFA.
}
\details{
See \code{vignette("ANALYTES_STATS")} for details.
}
\examples{
names(ANALYTES_EMM$timewise) # clinical analytes

str(ANALYTES_EMM$timewise)

# Print one of the emmGrid objects
ANALYTES_EMM$timewise$Glucagon

}
\seealso{
\code{\link{ANALYTES}}, \code{\link{ANALYTES_STATS}},
  \code{\link[emmeans]{emmeans}}
}
\keyword{datasets}
