% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ANALYTES_STATS}
\alias{ANALYTES_STATS}
\title{Statistical analyses of clinical analytes in plasma}
\format{
A list of \code{data.frame} objects with names "timewise"
  (trained vs. SED comparisons) and "MvF" (male vs. female comparisons).
}
\usage{
ANALYTES_STATS
}
\description{
Statistical analyses of several clinical analytes described in
  \code{\link{ANALYTES}}.
}
\details{
See \code{vignette("ANALYTES_STATS")} for details.
}
\examples{
head(ANALYTES_STATS$timewise) # trained vs. SED comparisons
head(ANALYTES_STATS$MvF) # male vs. female comparisons

}
\seealso{
\code{\link{ANALYTES}}, \code{\link{ANALYTES_EMM}}
}
\keyword{datasets}
