% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO_FGSEA}
\alias{GO_FGSEA}
\alias{TRNSCRPT_FGSEA}
\alias{PROT_FGSEA}
\title{Gene Ontology FGSEA results}
\format{
An object of class \code{list} of length 3.
}
\usage{
TRNSCRPT_FGSEA # transcriptomics
PROT_FGSEA     # proteomics
}
\description{
Gene Ontology FGSEA results
}
\keyword{datasets}
