% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m_phospho}
\alias{m_phospho}
\title{scWAT Phospho-proteomics MSnSet}
\format{
An \code{\link[MSnbase]{MSnSet-class}} object.
}
\usage{
data("m_phospho")
}
\description{
An MSnSet object containing the subcutaneous white adipose
  tissue (scWAT) phospho-proteomics data provided by the MoTrPAC Bioinformatics
  Center.
}
\keyword{datasets}
