% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m_prot}
\alias{m_prot}
\title{scWAT Proteomics MSnSet}
\format{
An \code{\link[MSnbase]{MSnSet-class}} object.
}
\usage{
data(m_prot)
}
\description{
An MSnSet object containing the subcutaneous white adipose
  tissue (scWAT) proteomics data provided by the MoTrPAC Bioinformatics
  Center.
}
\author{
Tyler Sagendorf
}
\keyword{datasets}
