% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m_trnscrpt}
\alias{m_trnscrpt}
\title{scWAT Transcriptomics MSnSet}
\format{
An \code{\link[MSnbase]{MSnSet-class}} object with 16443 features and
  48 samples.

  # exprs

  `exprs(m_trnscrpt)` is a count matrix that has been filtered to remove 321
  low-count features with \code{\link[edgeR]{filterByExpr}}.

  # fData

  `fData(m_trnscrpt)` is a \code{\link[base]{data.frame}} with 16443 rows and
  3 variables:
  \describe{
    \item{entrez_id}{character list; the Entrez gene(s) associated with each
    feature (Ensembl gene).}
    \item{gene_symbol}{character list; the gene symbol(s) associated with
    each feature.}
    \item{n}{integer; number of genes annotated to each feature.}
  }

  Prior to collapsing the `entrez_id` and `gene_symbol` columns into lists,
  any rows where `gene_symbol` began with "LOC" were removed unless that
  would remove entire features. This reduces the impact of the one-to-many
  feature-to-gene mapping on FGSEA results.

  # pData

  `pData(m_trnscrpt)` is a \code{\link[base]{data.frame}} with 48 rows
  (samples) and 9 variables:
  \describe{
    \item{viallabel}{character; unique sample identifier.}
    \item{pct_globin}{numeric; percent of reads mapping to globin.}
    \item{rin}{numeric; RNA integrity number (RIN).}
    \item{pct_umi_dup}{numeric; percent of PCR duplicates as quantified with Unique Molecular Identifiers (UMIs).}
    \item{median_5_3_bias}{numeric; median 5'-3' bias.}
    \item{sex}{character; "M" for males, "F" for females.}
    \item{timepoint}{character; "SED" for sedentary control animals. Otherwise, the weeks of exercise training that the animals underwent ("1W", "2W", "4W", "8W").}
    \item{exp_group}{character; the experimental group (each combination of `sex` and `timepoint`).}
    \item{bid}{character; Unique 5 digit numeric identifier of all samples collected for an acute test/sample collection period. All samples collected during that period will have the same BID. Same as first 5 digits of `viallabel`.}
  }
}
\usage{
data("m_trnscrpt")
}
\description{
An MSnSet object containing the subcutaneous white adipose
  tissue (scWAT) transcriptomics data provided by the MoTrPAC Bioinformatics
  Center.
}
\details{
Two outlier samples ("90423", "90410") were removed (48 remain). Then,
 `pct_globin`, `rin`, `pct_umi_dup`, and `median_5_3_bias` were mean-imputed,
 centered, and scaled.
}
\keyword{datasets}
