% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METAB_DA}
\alias{METAB_DA}
\title{Metabolomics differential analysis}
\format{
An object of class \code{list} of length 3.
}
\usage{
METAB_DA
}
\description{
Metabolomics differential analysis
}
\keyword{datasets}
