% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METAB_EXP}
\alias{METAB_EXP}
\title{Metabolomics ExpressionSet}
\format{
Object of class
  \code{\link[Biobase:ExpressionSet-class]{Biobase::ExpressionSet}} with 1063
  features and 50 samples.
}
\usage{
METAB_EXP
}
\description{
An object of class
  \code{\link[Biobase:ExpressionSet-class]{Biobase::ExpressionSet}}
  containing log\eqn{_2}-transformed metabolite data for differential
  analysis and WGCNA.
}
\examples{
library(Biobase)

METAB_EXP                   # summary
exprs(METAB_EXP)[1:10, 1:5] # assayData (first 10 rows and 5 columns)
head(fData(METAB_EXP))      # featureData
head(pData(METAB_EXP))      # phenoData

}
\references{
Fahy, E., & Subramaniam, S. (2020). RefMet: A reference
  nomenclature for metabolomics. \emph{Nature Methods, 17}(12), 1173–1174.
  \url{https://doi.org/10.1038/s41592-020-01009-y}
}
\seealso{
\code{\link[Biobase]{ExpressionSet-class}},
  \code{\link[MotrpacRatTraining6moData]{METAB_FEATURE_ID_MAP}},
  \code{\link[MotrpacRatTraining6moData]{METAB_NORM_DATA_NESTED}},
  \href{https://www.metabolomicsworkbench.org/databases/refmet/index.php}{Metabolomics
  Workbench RefMet Database}
}
\keyword{datasets}
