% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METAB_MSNSET}
\alias{METAB_MSNSET}
\title{Metabolomics MSnSet}
\format{
Object of class \code{\link[MSnbase:MSnSet-class]{MSnbase::MSnSet}}
  with 1063 features and 50 samples.

  \describe{
    \item{\bold{assayData}}{
      Object of class \code{matrix} with 1063 rows and 50 columns. See
      \link[MotrpacRatTraining6moData]{METAB_NORM_DATA_NESTED} for
      more details.
    }
    \item{\bold{featureData}}{
      Object of class \code{data.frame} with 1063 rows and 13 variables
      containing additional feature information:

      \describe{
         \item{\code{feature_ID}}{character; metabolite RefMet identifier.}
         \item{\code{dataset}}{character; metabolomics platform in which the
         feature was measured.}
         \item{\code{name_in_figures}}{character; alternative feature
         identifier primarily used in visualizations.}
         \item{\code{refmet_super_class}}{character; RefMet chemical super
         class.}
         \item{\code{refmet_main_class}}{character; RefMet chemical main
         class.}
         \item{\code{refmet_sub_class}}{character; RefMet chemical sub
         class.}
         \item{\code{lipid_class}}{character; (lipids only) the class of the
         lipid. There are 43 unique classes.}
         \item{\code{chain_length}}{integer; (lipids only) total number of
         carbons that comprise the fatty acid chains.}
         \item{\code{double_bond}}{integer; total number of C=C double bonds
         in the fatty acid chains.}
         \item{\code{rt}}{numeric; retention time. Used to assign
         \code{lipid_class}.}
         \item{\code{mz}}{numeric; mass-to-charge ratio (m/z).}
         \item{\code{neutral_mass}}{numeric; neutral mass (g/mol).}
         \item{\code{formula}}{character; chemical formula.}
      }
    }
    \item{\bold{phenoData}}{
      Object of class \code{data.frame} with 50 rows and 6 variables
      containing additional sample information:

      \describe{
        \item{\code{pid}}{integer; randomly generated 8-digit identifier used
        in linkage to phenotypic data. All samples from the same animal have
        the same PID.}
        \item{\code{bid}}{integer; unique 5 digit identifier of all
        samples collected for an acute test/sample collection period. All
        samples collected during that period will have the same BID.}
        \item{\code{labelid}}{integer; unique 11 digit specimen label
        identifier, originating at the collection site, that provides
        a link to specimen processing and is used for shipments to the
        biorepository. Same as \code{viallabel} only in instances
        where aliquots are not further processed at the biorepository.}
        \item{\code{sex}}{factor; the sex of the rat with levels "Female" and
        "Male".}
        \item{\code{timepoint}}{factor; exercise training group. Either "SED"
        (sedentary) or the number of weeks of training ("1W", "2W", "4W",
        "8W").}
        \item{\code{exp_group}}{factor; experimental group. Unique
        combination of \code{sex} (first letter) and \code{timepoint}.}
      }
    }
  }
}
\usage{
METAB_MSNSET
}
\description{
An object of class
  \code{\link[MSnbase:MSnSet-class]{MSnbase::MSnSet}} containing
  log\eqn{_2}-transformed metabolite data for differential analysis and
  WGCNA.
}
\examples{
library(MSnbase)
METAB_MSNSET                   # summary
exprs(METAB_MSNSET)[1:10, 1:5] # assayData (first 10 rows and 5 columns)
head(fData(METAB_MSNSET))      # featureData
head(pData(METAB_MSNSET))      # phenoData

}
\references{
Fahy, E., & Subramaniam, S. (2020). RefMet: A reference
  nomenclature for metabolomics. \emph{Nature Methods, 17}(12), 1173--1174.
  \url{https://doi.org/10.1038/s41592-020-01009-y}
}
\seealso{
\link[MSnbase]{MSnSet-class},
  \link[MotrpacRatTraining6moData]{METAB_FEATURE_ID_MAP},
  \link[MotrpacRatTraining6moData]{METAB_NORM_DATA_NESTED},
  \href{https://www.metabolomicsworkbench.org/databases/refmet/index.php}{Metabolomics
  Workbench RefMet Database}
}
\keyword{datasets}
