% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MITOCARTA_HS}
\alias{MITOCARTA_HS}
\title{Human MitoCarta3.0 Pathways}
\format{
A \code{data.frame} with 149 rows (pathways) and 3 variables:

\describe{
  \item{pathway}{character; pathway description.}
  \item{hierarchy}{character; pathway hierarchy.}
  \item{human_genes}{list; character list of human genes in each pathway.}
}
}
\source{
The "HumanMitoCarta3.0.xls" file was downloaded from
  \url{https://www.broadinstitute.org/files/shared/metabolism/mitocarta/human.mitocarta3.0.html}.
  Columns 2–4 were extracted from the "C MitoPathways" sheet and renamed.
}
\usage{
MITOCARTA_HS
}
\description{
Pathways from the human MitoCarta3.0 database.
}
\references{
Rath, S., Sharma, R., Gupta, R., Ast, T., Chan, C., Durham, T.
  J., Goodman, R. P., Grabarek, Z., Haas, M. E., Hung, W. H. W., Joshi, P.
  R., Jourdain, A. A., Kim, S. H., Kotrys, A. V., Lam, S. S., McCoy, J. G.,
  Meisel, J. D., Miranda, M., Panda, A., Patgiri, A., … Mootha, V. K. (2021).
  MitoCarta3.0: an updated mitochondrial proteome now with sub-organelle
  localization and pathway annotations. \emph{Nucleic acids research,
  49}(D1), D1541–D1547. \url{https://doi.org/10.1093/nar/gkaa1011}
}
\seealso{
\url{https://personal.broadinstitute.org/scalvo/MitoCarta3.0/human.mitopathways3.0.html}
}
\keyword{datasets}
