% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotrpacRatTraining6moWATData-package.R
\docType{package}
\name{MotrpacRatTraining6moWATData-package}
\alias{MotrpacRatTraining6moWATData}
\alias{MotrpacRatTraining6moWATData-package}
\title{MotrpacRatTraining6moWATData: Data and Analysis Results for the MoTrPAC PASS1B Rat White Adipose Tissue Companion Paper}
\description{
Contains data and code necessary to reproduce analyses that appear in "Sexual dimorphism and the multi-omic response to exercise training in rat subcutaneous white adipose tissue".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://motrpac.github.io/MotrpacRatTraining6moWATData/}
  \item \url{https://github.com/MoTrPAC/MotrpacRatTraining6moWATData/}
}

}
\author{
\strong{Maintainer}: Tyler J. Sagendorf \email{tyler.sagendorf@pnnl.gov} (\href{https://orcid.org/0000-0003-1552-4870}{ORCID})

Authors:
\itemize{
  \item Zhenxin Hou (\href{https://orcid.org/0000-0002-1301-0799}{ORCID}) (Provided WGCNA-related code, triacylglycerol abundance data, and both notes and data regarding the redundant metabolite filtering process.)
}

Other contributors:
\itemize{
  \item Christopher B. Newgard (Provided clinical analyte data.) [data contributor]
  \item Michael J. Muehlbauer (Provided clinical analyte data.) [data contributor]
  \item Pasquale Nigro (\href{https://orcid.org/0000-0002-2918-8356}{ORCID}) (Provided histology data.) [data contributor]
  \item Malene E. Lindholm (\href{https://orcid.org/0000-0002-5763-7833}{ORCID}) (Provided mtDNA qPCR results.) [data contributor]
}

}
\keyword{internal}
