% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHENO_WAT}
\alias{PHENO_WAT}
\title{scWAT Phenotype Data}
\format{
An object of class \code{data.frame} with 92 rows and 21 columns.
}
\usage{
PHENO_WAT
}
\description{
Phenotype data for scWAT samples from the MoTrPAC endurance
  exercise training study in 6-month-old-rats: NMR body composition and
  VO\eqn{_2}max testing results.
}
\examples{
head(PHENO_WAT)

}
\keyword{datasets}
