% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHENO_WAT_STATS}
\alias{PHENO_WAT_STATS}
\title{Statistical analyses of scWAT phenotype data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27 rows and 16 columns.
}
\usage{
PHENO_WAT_STATS
}
\description{
Analyses of phenotypic data from \code{\link{PHENO_WAT}}. See
  \code{vignette("PHENO_WAT_STATS")} for details.
}
\examples{
head(PHENO_WAT_STATS)

}
\keyword{datasets}
