% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHOSPHO_DA}
\alias{PHOSPHO_DA}
\title{Phosphoproteomics differential analysis}
\format{
A named list of `data.frames`:

    "trained_vs_SED": all trained timepoints compared to their sex-matched
    sedentary control group. Example: F_1W - F_SED. Total of 8 contrasts and
    30,304*8 = 242,432 rows.

    "MvF": sedentary males compared to sedentary females. Only 1 contrast,
    30,304 rows.

    "DD": comparisons of the male and female training responses. Example:
    (M_8W - M_SED) - (F_8W - F_SED). Total of 4 contrasts and 30,304*4 =
    121,216 rows.

    Each data.frame has 15 columns, most of which are from
    \code{\link[limma]{topTable}}:

    \describe{
      \item{feature_ID}{character; combination of RefSeq protein identifier
      and phosphorylation sites.}
      \item{gene_symbol}{character; gene symbol.}
      \item{entrez_gene}{numeric; Entrez gene identifier.}
      \item{site}{character; phosphorylation site(s) extracted from the
      "feature_ID" column and reformatted slightly. Multiple instances of
      phosphorylation on the same protein are separated by semicolons (;).}
      \item{human_feature_ID}{character; combination of human UniProt protein
      identifier and phosphorylation sites. Similar to the "feature_ID"
      column.}
      \item{human_uniprot}{character; human UniProt protein identifiers.}
      \item{human_site}{character; equivalent human phosphorylation site.}
      \item{logFC}{numeric; difference in the weighted mean \eqn{log_2} relative
      abundances of the groups specified in the "contrast" column.}
      \item{AveExpr}{numeric; mean \eqn{log_2} relative abundance of each
      protein.}
      \item{t}{numeric; LIMMA moderated t-statistic.}
      \item{P.Value}{numeric; p-value.}
      \item{adj.P.Val}{numeric; Benjamini-Hochberg adjusted p-value. P-values
      are adjusted across all contrasts.}
      \item{feature}{character; same as "feature_ID".}
      \item{contrast}{factor; the contrast(s) being tested.}
      \item{se}{numeric; standard error (SE) of the "logFC" column.}
    }
}
\usage{
PHOSPHO_DA
}
\description{
Differential analysis results of 30,304 phosphorylation sites
  measured in rat scWAT.
}
\keyword{datasets}
