% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHOSPHO_EXP}
\alias{PHOSPHO_EXP}
\title{Phosphoproteomics ExpressionSet}
\format{
Object of class
  \code{\link[Biobase:ExpressionSet-class]{Biobase::ExpressionSet}} with
  30304 features and 60 samples.
}
\usage{
PHOSPHO_EXP
}
\description{
An object of class
  \code{\link[Biobase:ExpressionSet-class]{Biobase::ExpressionSet}}
  containing median-MAD-normalized log\eqn{_2}-transformed protein
  phosphorylation data.
}
\details{
\code{\link[MotrpacRatTraining6moData]{RAT_TO_HUMAN_PHOSPHO}} for
  details of mapping between rat and human phosphorylation sites.

  See \code{\link[MotrpacRatTraining6moData]{FEATURE_TO_GENE}} for details of
  feature-to-gene mapping.
}
\examples{
library(Biobase)

PHOSPHO_EXP                   # summary
exprs(PHOSPHO_EXP)[1:10, 1:5] # assayData (first 10 rows and 5 columns)
head(fData(PHOSPHO_EXP))      # featureData
head(pData(PHOSPHO_EXP))      # phenoData

}
\seealso{
\code{\link[Biobase]{ExpressionSet-class}},
  \code{\link[MotrpacRatTraining6moData]{RAT_TO_HUMAN_PHOSPHO}},
  \code{\link[MotrpacRatTraining6moData]{FEATURE_TO_GENE}}
}
\keyword{datasets}
