% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHOSPHO_KSEA}
\alias{PHOSPHO_KSEA}
\title{Phosphoproteomics KSEA}
\format{
An object of class \code{list} of length 3.
}
\usage{
PHOSPHO_KSEA
}
\description{
Phosphoproteomics KSEA
}
\keyword{datasets}
