% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHOSPHO_MSNSET}
\alias{PHOSPHO_MSNSET}
\title{Phosphoproteomics MSnSet}
\format{
An \code{\link[MSnbase]{MSnSet-class}} object with 9964 features and
60 samples.
\subsection{\code{assayData}}{

A \code{matrix} with 30304 rows and 60 columns.
}

\subsection{\code{featureData}}{

A \code{data.frame} with 30304 rows and 7 variables containing additional
feature information:
\describe{
\item{feature_ID}{character; }
\item{gene_symbol}{character list; the gene symbol associated with each
protein.}
\item{entrez_gene}{numeric list; the Entrez gene identifier associated
with each protein.}
\item{site}{character; }
\item{human_feature_ID}{character; }
\item{human_uniprot}{character; }
\item{human_site}{character; }
}
}

\subsection{\code{phenoData}}{

A \code{data.frame} with 60 rows and 7 variables containing sample information:
\describe{
\item{pid}{numeric; }
\item{bid}{numeric; }
\item{labelid}{numeric; }
\item{viallabel}{character; }
\item{sex}{factor; the sex of the rat with 2 levels "Female" and "Male".}
\item{timepoint}{factor; exercise training group. Either "SED"
(sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").}
\item{exp_group}{factor; unique combination of \code{sex} (first letter) and
\code{timepoint}.}
}
}
}
\usage{
PHOSPHO_MSNSET
}
\description{
An \code{\link[MSnbase]{MSnSet-class}} object containing
phosphoproteomics data from rat WAT provided by MoTrPAC.
}
\examples{
PHOSPHO_MSNSET # summary
library(MSnbase)
exprs(PHOSPHO_MSNSET)[1:5, 1:5] # assayData
head(fData(PHOSPHO_MSNSET))     # featureData
head(pData(PHOSPHO_MSNSET))     # phenoData

}
\keyword{datasets}
