% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PLASMA_ANALYTE_STATS}
\alias{PLASMA_ANALYTE_STATS}
\title{Plasma analyte statistical analyses}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 66 rows and 13 columns.
}
\usage{
PLASMA_ANALYTE_STATS
}
\description{
Statistical analyses of most clinical analytes described in
  \code{\link[MotrpacRatTraining6moWATData]{PLASMA_ANALYTES}}.
}
\keyword{datasets}
