% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PLASMA_ANALYTES}
\alias{PLASMA_ANALYTES}
\title{Plasma clinical analytes}
\format{
An object of class \code{data.frame} with 146 rows and 21 columns.

\describe{
  \item{\code{plate}}{integer; 96-well plate identifier (1--4).}
  \item{\code{runseq}}{integer; run order (1--146).}
  \item{\code{platepos}}{integer; position in each 96-well plate.}
  \item{\code{pid}}{integer; randomly generated 8-digit identifier used in
  linkage to phenotypic data. All samples from the same animal have the same
  PID.}
  \item{\code{bid}}{integer; unique 5 digit identifier of all samples
  collected for an acute test/sample collection period. All samples collected
  during that period will have the same BID.}
  \item{\code{labelid}}{integer; unique 11 digit specimen label identifier,
  originating at the collection site, that provides a link to specimen
  processing and is used for shipments to the biorepository. Same as
  \code{viallabel} only in instances where aliquots are not further processed
  at the biorepository.}
  \item{\code{omics_subset}}{logical; whether the sample was selected for
  -omics analysis.}
  \item{\code{sex}}{factor; the sex of the rat with levels "Female" and
  "Male".}
  \item{\code{timepoint}}{factor; exercise training group. Either "SED"
  (sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").}
  \item{\code{total_ketones}}{numeric; total ketone bodies (\eqn{\mu}mol/L).}
  \item{\code{nefa}}{numeric; non-esterified fatty acids (NEFA) (mmol/L).}
  \item{\code{lactate}}{numeric; lactate (mmol/L).}
  \item{\code{glucose}}{numeric; glucose (mg/dL).}
  \item{\code{glycerol}}{numeric; glycerol (mg/dL).}
  \item{\code{glucagon}}{numeric; glucagon (pmol).}
  \item{\code{insulin}}{numeric; insulin (pg/mL).}
  \item{\code{insulin_iu}}{numeric; insulin (\eqn{\mu}IU/mL) using the
  conversion 1 pg/mL = 0.023 \eqn{\mu}IU/mL provided in the package insert.}
  \item{\code{leptin}}{numeric; leptin (pg/mL).}
  \item{\code{corticosterone}}{numeric; corticosterone (ng/mL).}
  \item{\code{ins_gcg_ratio}}{numeric; molar insulin/glucagon ratio.
  Calculated as (insulin (pg/mL) / 5.804 (Da)) / glucagon (pmol).}
  \item{\code{homa_ir}}{numeric; Homeostatic Model Assessment for Insulin
  Resistance (HOMA-IR). Insulin (\eqn{\mu}IU/mL) x glucose (mg/dL) / 405.}
}
}
\usage{
PLASMA_ANALYTES
}
\description{
A set of 9 common clinical analytes measured in plasma.
}
\details{
\bold{Fujifilm Wako (Osaka, Japan):}
\itemize{
  \item Total ketones: \emph{415-73301, 411-73401}
  \item NEFA: \emph{995-34791, 999-34691, 993-35191, 991-34891}
}

\bold{Beckman (Brea, CA):}
\itemize{
  \item Lactate: \emph{A95550}
  \item Glucose: \emph{B24985}
  \item Glycerol: \emph{445850}
}

\bold{Meso Scale Discovery (Rockville, MD):}
\itemize{
  \item Glucagon: \emph{K1535YK}
  \item Insulin, Leptin: \emph{K15158C (multiplex assay)}
}

\bold{Alpco (Salem, NH):}
\itemize{
  \item Corticosterone: \emph{55-CORMS-E01}
}
}
\keyword{datasets}
