% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upset.R
\name{plot_upset}
\alias{plot_upset}
\title{Create Annotated UpSet Plot}
\usage{
plot_upset(
  ...,
  mode = c("distinct", "intersect", "union"),
  top_n_comb = 10,
  scale_set_bars = FALSE,
  row_labels = character(0),
  row_names_gp = gpar(fontsize = 6.5),
  annotation_name_gp = gpar(fontsize = 6.5, fontface = "bold"),
  heatmap_args = list(),
  cell_height = unit(4, "mm"),
  cell_width = cell_height,
  extend = 0.2,
  rot = 45,
  hjust = 0,
  vjust = 0,
  bar_label_gp = gpar(fontsize = 6),
  filename = character(0),
  height = 2.5,
  width = 2.7,
  save_args = list()
)
}
\arguments{
\item{...}{The input sets. If it is represented as a single variable, it should be a matrix/data frame or a list. If it is multiple variables, it should be name-value pairs, see Input section for explanation.}

\item{mode}{The mode for forming the combination set, see Mode section.}

\item{top_n_comb}{integer; number of largest intersections to display.
Default is 10.}

\item{scale_set_bars}{logical;}

\item{row_labels}{character; labels for each set (row of UpSet plot).
Defaults to rownames of \code{\link[ComplexHeatmap]{make_comb_mat}} output.}

\item{row_names_gp}{graphical parameters specified by
\code{\link[grid]{gpar}}. Modifies \code{row_labels}.}

\item{annotation_name_gp}{graphical parameters specified by
\code{\link[grid]{gpar}}. Passed to
\code{\link[ComplexHeatmap]{HeatmapAnnotation}}. Modifies the barplot
titles.}

\item{heatmap_args}{list; additional arguments passed to
\code{\link[ComplexHeatmap]{Heatmap}}.}

\item{cell_height}{\code{unit} object; the height of each heatmap cell. Default is
\code{unit(4, "mm")}.}

\item{cell_width}{same as \code{cell_height}.}

\item{extend}{numeric;}

\item{rot}{numeric; angle in [0,360] to rotate set intersection labels.}

\item{hjust}{numeric;}

\item{vjust}{numeric;}

\item{bar_label_gp}{graphical parameters specified by
\code{\link[grid]{gpar}}. Passed to
\code{\link[ComplexHeatmap]{HeatmapAnnotation}}. Modifies the annotation
labels.}

\item{filename}{character; the filename used to save the heatmap. If
\code{character(0)} (default), the heatmap will be displayed instead.}

\item{height}{numeric; height of the file in \code{units}.}

\item{width}{same as \code{height}.}

\item{save_args}{list; additional arguments passed to the graphics device.
See \code{\link[grDevices]{png}} for options.}
}
\description{
Create an UpSet plot annotated with intersection and set sizes.
Essentially a modified version of \code{\link[ComplexHeatmap]{UpSet}}.
}
\details{
See \code{\link[ComplexHeatmap]{UpSet}} for details and sections
mentioned in Arguments.
}
\references{
Lex, A., Gehlenborg, N., Strobelt, H., Vuillemot, R., & Pfister,
H. (2014). UpSet: Visualization of Intersecting Sets. \emph{IEEE transactions on
visualization and computer graphics, 20}(12), 1983--1992.
\url{https://doi.org/10.1109/TVCG.2014.2346248}
}
