% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Create Volcano Plot}
\usage{
plot_volcano(
  x,
  logFC = "logFC",
  pval = "adj.P.Val",
  pval_cutoff,
  label,
  scale = 2.5
)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame} containing differential analysis results.}

\item{logFC}{character; name of a column in \code{x} containing log2 fold-changes.
Default is "logFC".}

\item{pval}{character; name of a column in \code{x} contain (adjusted) p-values.
Default is "adj.P.Val".}

\item{pval_cutoff}{numeric; cutoff for p-values to be considered significant.
If provided, a horizontal dashed line will be added to the plot.}

\item{label}{character; name of a column in \code{x} used to label the points. Any
points that should not be labelled should have a missing label.}

\item{scale}{numeric; scaling factor used when saving the plot. Default is
2.5.}
}
\description{
Create a volcano plot with optional labels.
}
