% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROT_DA}
\alias{PROT_DA}
\title{Proteomics differential analysis}
\format{
A named list of `data.frames`:

    "trained_vs_SED": all trained timepoints compared to their sex-matched
    sedentary control group. Example: F_1W - F_SED. Total of 8 contrasts and
    9964*8 = 79,712 rows.

    "MvF": sedentary males compared to sedentary females. Only 1 contrast,
    9964 rows.

    "DD": comparisons of the male and female training responses. Example:
    (M_8W - M_SED) - (F_8W - F_SED). Total of 4 contrasts and 9964*4 = 39,856
    rows.

    Each data.frame has 11 columns, most of which are from
    \code{\link[limma]{topTable}}:

    \describe{
      \item{feature_ID}{character; Reference Sequence (RefSeq) protein
      identifier.}
      \item{gene_symbol}{character; gene symbol.}
      \item{entrez_gene}{numeric; Entrez gene identifier.}
      \item{logFC}{numeric; difference in the weighted mean \eqn{log_2} relative
      abundances of the groups specified in the "contrast" column.}
      \item{AveExpr}{numeric; mean \eqn{log_2} relative abundance of each
      protein.}
      \item{t}{numeric; LIMMA moderated t-statistic.}
      \item{P.Value}{numeric; p-value.}
      \item{adj.P.Val}{numeric; Benjamini-Hochberg adjusted p-value. P-values
      are adjusted across all contrasts.}
      \item{feature}{character; same as "feature_ID".}
      \item{contrast}{factor; the contrast(s) being tested.}
      \item{se}{numeric; standard error (SE) of the "logFC" column.}
    }
}
\usage{
PROT_DA
}
\description{
Differential analysis results of 9964 proteins measured in rat
  scWAT.
}
\keyword{datasets}
