% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROT_EXP}
\alias{PROT_EXP}
\title{Proteomics ExpressionSet}
\format{
Object of class
  \code{\link[Biobase:ExpressionSet-class]{Biobase::ExpressionSet}} with 9964
  features and 60 samples.
}
\usage{
PROT_EXP
}
\description{
An object of class
  \code{\link[Biobase:ExpressionSet-class]{Biobase::ExpressionSet}}
  containing median-MAD-normalized log\eqn{_2}-transformed protein expression
  data for differential analysis and WGCNA.
}
\details{
See \code{\link[MotrpacRatTraining6moData]{FEATURE_TO_GENE}} for
  details of feature-to-gene mapping.
}
\examples{
library(Biobase)

PROT_EXP                   # display summary
exprs(PROT_EXP)[1:10, 1:5] # assayData (first 10 rows and 5 columns)
head(fData(PROT_EXP))      # featureData
head(pData(PROT_EXP))      # phenoData

}
\seealso{
\code{\link[Biobase]{ExpressionSet-class}},
  \code{\link[MotrpacRatTraining6moData]{PROT_WATSC_NORM_DATA}},
  \code{\link[MotrpacRatTraining6moData]{FEATURE_TO_GENE}}
}
\keyword{datasets}
