% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROT_FGSEA}
\alias{PROT_FGSEA}
\title{Proteomics Gene Ontology FGSEA}
\format{
An object of class \code{list} of length 3.
}
\usage{
PROT_FGSEA
}
\description{
Proteomics Gene Ontology FGSEA
}
\keyword{datasets}
