% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROT_MSNSET}
\alias{PROT_MSNSET}
\title{Proteomics MSnSet}
\format{
An \code{\link[MSnbase]{MSnSet-class}} object with 9964 features and
60 samples.
\subsection{\code{assayData}}{

A \code{matrix} with 9964 rows and 60 columns.
}

\subsection{\code{featureData}}{

A \code{data.frame} with 9964 rows and 3 variables containing additional feature
information:
\describe{
\item{feature_ID}{character; Reference Sequence (RefSeq) protein
identifier.}
\item{gene_symbol}{character list; the gene symbol associated with each
protein.}
\item{entrez_gene}{numeric list; the Entrez gene identifier associated
with each protein.}
}
}

\subsection{\code{phenoData}}{

A \code{data.frame} with 60 rows and 7 variables containing sample information:
\describe{
\item{pid}{numeric; }
\item{bid}{numeric; }
\item{labelid}{numeric; }
\item{viallabel}{character; }
\item{sex}{factor; the sex of the rat with 2 levels "Female" and "Male".}
\item{timepoint}{factor; exercise training group. Either "SED"
(sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").}
\item{exp_group}{factor; unique combination of \code{sex} (first letter) and
\code{timepoint}.}
}
}
}
\usage{
PROT_MSNSET
}
\description{
An \code{\link[MSnbase]{MSnSet-class}} object containing
proteomics data from rat WAT provided by MoTrPAC.
}
\examples{
PROT_MSNSET # summary
library(MSnbase)
exprs(PROT_MSNSET)[1:6, 1:5] # assayData
head(fData(PROT_MSNSET))     # featureData
head(pData(PROT_MSNSET))     # phenoData

}
\keyword{datasets}
