% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_hist.R
\name{pval_hist}
\alias{pval_hist}
\title{Create p-value Histograms}
\usage{
pval_hist(x, pval = "adj.P.Val", ylims = c(0, NA), scale = 2.5)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame} containing differential analysis results.}

\item{pval}{character; name of a column in \code{x} containing (adjusted)
p-values. Default is "adj.P.Val".}

\item{ylims}{numeric vector of length 2. Limits of y-axis.}

\item{scale}{numeric; scaling factor used when saving the plot. Default is
2.5.}
}
\description{
Create p-value histograms with
  \code{\link[ggplot2]{ggplot2-package}}.
}
