% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_outer.R
\name{round_outer}
\alias{round_outer}
\title{Round to Outer Limits}
\usage{
round_outer(x, nearest = 1)
}
\arguments{
\item{x}{a numeric vector.}

\item{nearest}{numeric; the range of \code{x} will be rounded out to the
value specified by \code{nearest}. Default is 1.}
}
\value{
A numeric vector of length 2 specifying the range of values after
  rounding outward to the nearest value provided by `nearest`.
}
\description{
Round limits out to nearest value.
}
\examples{
set.seed(0)
x <- runif(5, min = -10, max = 10)
x

round_outer(x) # nearest whole number
round_outer(x, nearest = 0.1) # nearest tenth
}
\author{
Tyler Sagendorf
}
