% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRNSCRPT_DA}
\alias{TRNSCRPT_DA}
\title{Transcriptomics differential analysis}
\format{
A named list of `data.frames`:

    "trained_vs_SED": all trained timepoints compared to their sex-matched
    sedentary control group. Example: F_1W - F_SED. Total of 8 contrasts and
    16,547*8 = 132,376 rows.

    "MvF": sedentary males compared to sedentary females. Only 1 contrast,
    16,547 rows.

    "DD": comparisons of the male and female training responses. Example:
    (M_8W - M_SED) - (F_8W - F_SED). Total of 4 contrasts and 16,547*4 =
    66,188 rows.

    Each data.frame has 11 columns, most of which are from
    \code{\link[limma]{topTable}}:

    \describe{
      \item{feature_ID}{character; transcript gene identifier.}
      \item{gene_symbol}{character; gene symbol. For 195 of the transcripts,
      there is a one-to-many mapping between transcripts and genes. Multiple
      genes are separated by semicolons (;).}
      \item{entrez_gene}{numeric; Entrez gene identifier. For 180 of the
      transcripts, there is a one-to-many mapping between transcripts and
      genes. Multiple genes are separated by semicolons (;).}
      \item{logFC}{numeric; difference in the weighted mean \eqn{log_2} relative
      abundances of the groups specified in the "contrast" column.}
      \item{AveExpr}{numeric; mean \eqn{log_2} relative abundance of each
      protein.}
      \item{t}{numeric; LIMMA moderated t-statistic.}
      \item{P.Value}{numeric; p-value.}
      \item{adj.P.Val}{numeric; Benjamini-Hochberg adjusted p-value. P-values
      are adjusted across all contrasts.}
      \item{feature}{character; same as "feature_ID".}
      \item{contrast}{factor; the contrast(s) being tested.}
      \item{se}{numeric; standard error (SE) of the "logFC" column.}
    }
}
\usage{
TRNSCRPT_DA
}
\description{
Differential analysis results of 16,547 transcripts measured in
  rat scWAT.
}
\keyword{datasets}
