% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRNSCRPT_EXP}
\alias{TRNSCRPT_EXP}
\title{Transcriptomics ExpressionSet}
\format{
Object of class
  \code{\link[Biobase:ExpressionSet-class]{Biobase::ExpressionSet}} with
  16404 features and 48 samples.
}
\usage{
TRNSCRPT_EXP
}
\description{
An object of class
  \code{\link[Biobase:ExpressionSet-class]{Biobase::ExpressionSet}}
  containing transcriptomics data for differential analysis and (once
  transformed to log\eqn{_2} counts per million reads) WGCNA.
}
\details{
The original 32883 transcripts were filtered with
  \code{\link[edgeR]{filterByExpr}} using \code{group =
  TRNSCRPT_EXP[["exp_group"]]}. Then, library sizes were TMM normalized
  with \code{\link[edgeR]{calcNormFactors}}. Two outlier samples (90423017005
  and 90410017005) were removed before variables \code{pct_globin},
  \code{rin}, \code{pct_umi_dup}, and \code{median_5_3_bias} were
  mean-imputed, centered, and scaled.

  See \code{\link[MotrpacRatTraining6moData]{OUTLIERS}} for sample outlier
  information.

  See \code{\link[MotrpacRatTraining6moData]{FEATURE_TO_GENE}} for details of
  feature-to-gene mapping.
}
\examples{
library(Biobase)

TRNSCRPT_EXP                   # summary
exprs(TRNSCRPT_EXP)[1:10, 1:5] # assayData (first 10 rows and 5 columns)
head(fData(TRNSCRPT_EXP))      # featureData
head(pData(TRNSCRPT_EXP))      # phenoData

}
\seealso{
\code{\link[Biobase]{ExpressionSet-class}},
  \code{\link[MotrpacRatTraining6moData]{TRNSCRPT_WATSC_RAW_COUNTS}},
  \code{\link[MotrpacRatTraining6moData]{FEATURE_TO_GENE}},
  \code{\link[edgeR]{DGEList}}, \code{\link[edgeR]{filterByExpr}},
  \code{\link[edgeR]{calcNormFactors}}
}
\keyword{datasets}
