% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRNSCRPT_FGSEA}
\alias{TRNSCRPT_FGSEA}
\title{Transcriptomics Gene Ontology FGSEA}
\format{
An object of class \code{list} of length 3.
}
\usage{
TRNSCRPT_FGSEA
}
\description{
Transcriptomics Gene Ontology FGSEA
}
\keyword{datasets}
