% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRNSCRPT_WGCNA}
\alias{TRNSCRPT_WGCNA}
\title{Transcriptomics WGCNA}
\format{
An object of class \code{list} of length 2.
}
\usage{
TRNSCRPT_WGCNA
}
\description{
Transcriptomics WGCNA
}
\keyword{datasets}
