% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WATSC_DA}
\alias{WATSC_DA}
\alias{PROT_DA}
\alias{PHOSPHO_DA}
\alias{TRNSCRPT_DA}
\alias{METAB_DA}
\title{Differential analysis results}
\format{
A named list of 3 \code{data.frame} objects:

  \describe{
    \item{"trained_vs_SED"}{All trained timepoints compared to their
    sex-matched sedentary control group. Example: F_1W - F_SED. Total of 8
    contrasts.}
    \item{"MvF_SED"}{Sedentary males compared to sedentary females. Only 1
    contrast.}
    \item{"MvF_exercise_response"}{Comparisons of the male and female
    training responses. Example: (M_8W - M_SED) - (F_8W - F_SED). Total of 4
    contrasts.}
  }

  In addition to all columns present in the \code{MSnbase::fData} tables,
  each \code{data.frame} contains the following 8 variables:

  \describe{
    \item{\code{logFC}}{numeric; difference in the weighted mean log\eqn{_2}
    values of the groups specified in the \code{contrast} column.}
    \item{\code{AveExpr}}{numeric; mean log\eqn{_2} relative abundance of
    each feature.}
    \item{\code{t}}{numeric; \acronym{LIMMA} moderated t-statistic.}
    \item{\code{P.Value}}{numeric; p-value.}
    \item{\code{adj.P.Val}}{numeric; Benjamini-Hochberg adjusted p-value.
    P-values are adjusted across all contrasts.}
    \item{\code{feature}}{character; uniquely defines each feature that was
    tested. Same as \code{feature_ID}.}
    \item{\code{contrast}}{factor; the contrast(s) being tested.}
    \item{\code{se}}{numeric; standard error (\acronym{SE}) of the
    \code{logFC} column.}
  }
}
\usage{
PROT_DA     # proteomics
PHOSPHO_DA  # phosphoproteomics
TRNSCRPT_DA # transcriptomics
METAB_DA    # metabolomics
}
\description{
Differential analysis of proteins, transcripts,
  phosphorylation sites, and metabolites measured in rat scWAT.
}
\details{
Differential analysis was performed with
  \code{\link[motrpacWAT:limma_full]{motrpacWAT::limma_full}} on the MSnSet
  objects (See Also section).
}
\examples{
## Number of differential features (-ome-wide FDR < 0.05)
# Convenience function
f1 <- function(x) {
  purrr::map(x, with, table(contrast, adj.P.Val < 0.05))
}

f1(PROT_DA)     # Proteins
f1(PHOSPHO_DA)  # Phosphosites
f1(TRNSCRPT_DA) # Transcripts
f1(METAB_DA)    # Metabolites

}
\seealso{
\code{\link[limma]{topTable}}, \code{\link[motrpacWAT]{limma_full}},
  \code{\link{PROT_MSNSET}}, \code{\link{PHOSPHO_MSNSET}},
  \code{\link{TRNSCRPT_MSNSET}}, \code{\link{METAB_MSNSET}}
}
\keyword{datasets}
