% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WATSC_WGCNA}
\alias{WATSC_WGCNA}
\alias{PROT_WGCNA}
\alias{TRNSCRPT_WGCNA}
\alias{METAB_WGCNA}
\title{WGCNA results}
\format{
An object of class \code{list} of length 2.
}
\usage{
PROT_WGCNA     # proteomics
TRNSCRPT_WGCNA # transcriptomics
METAB_WGCNA    # metabolomics
}
\description{
WGCNA results
}
\keyword{datasets}
