% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADIPOCYTE_SIZE}
\alias{ADIPOCYTE_SIZE}
\title{Measures of adipocyte size}
\format{
A \code{data.frame} with 55825 rows and 8 variables.
}
\usage{
ADIPOCYTE_SIZE
}
\description{
Measures of adipocyte size.
}
\keyword{datasets}
