% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METAB_FGSEA}
\alias{METAB_FGSEA}
\title{Metabolomics RefMet chemical subclass FGSEA}
\format{
An object of class \code{list} of length 3.
}
\usage{
METAB_FGSEA
}
\description{
Metabolomics RefMet chemical subclass FGSEA
}
\keyword{datasets}
