% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METAB_MSNSET}
\alias{METAB_MSNSET}
\title{Metabolomics MSnSet}
\format{
An \code{\link[MSnbase]{MSnSet-class}} object with 1063 non-redundant features and 50 samples.
}
\usage{
METAB_MSNSET
}
\description{
An MSnSet object containing the subcutaneous white adipose
  tissue (scWAT) metabolomics data provided by the MoTrPAC Bioinformatics
  Center.
}
\keyword{datasets}
