% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METAB_TG_CONC}
\alias{METAB_TG_CONC}
\title{Metabolomics triacylglyceride concentration}
\format{
An object of class \code{ExpressionSet} with 190 rows and 10 columns.
}
\usage{
METAB_TG_CONC
}
\description{
Metabolomics triacylglyceride concentration
}
\keyword{datasets}
