% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{module_ORA}
\alias{module_ORA}
\alias{PROT_MODULE_ORA}
\alias{TRNSCRPT_MODULE_ORA}
\alias{METAB_MODULE_ORA}
\title{Over-representation analysis of WGCNA modules}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2615 rows and 12 columns.
}
\usage{
PROT_MODULE_ORA     # proteomics
TRNSCRPT_MODULE_ORA # transcriptomics
METAB_MODULE_ORA    # metabolomics
}
\description{
Results of performing Gene Ontology (GO) over-representation
  analysis (Fisher Exact / Hypergeometric test) on the modules produced by
  weighted gene co-expression network analysis (WGCNA).
}
\keyword{datasets}
