% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PLASMA_ANALYTE_STATS}
\alias{PLASMA_ANALYTE_STATS}
\title{Statistical analyses of clinical analytes in plasma}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 44 rows and 17 columns.
}
\usage{
PLASMA_ANALYTE_STATS
}
\description{
Statistical analyses of most clinical analytes described in
  \code{\link{PLASMA_ANALYTES}}.
}
\keyword{datasets}
