% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Create Volcano Plot}
\usage{
plot_volcano(
  x,
  pval_cutoff = 0.05,
  scale = 2.5,
  colors = c("#3366ff", "darkred")
)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame} containing differential analysis results.}

\item{pval_cutoff}{numeric; cutoff for p-values to be considered significant. Adds a dashed horizontal line to the plot.}

\item{scale}{numeric; scaling factor used when saving the plot. Default is
2.5.}

\item{colors}{character; length 2 vector of negative and positive logFC colors. Only points with "adj.P.Val" < \code{pval_cutoff} will be colored.}
}
\value{
A \code{ggplot2} object.
}
\description{
Create a volcano plot.
}
