% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROT_MITOCARTA_FGSEA}
\alias{PROT_MITOCARTA_FGSEA}
\title{Proteomics MitoCarta FGSEA}
\format{
An object of class \code{list} of length 3.
}
\usage{
PROT_MITOCARTA_FGSEA
}
\description{
Proteomics MitoCarta FGSEA
}
\keyword{datasets}
