% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRNSCRPT_MSNSET}
\alias{TRNSCRPT_MSNSET}
\title{Transcriptomics MSnSet}
\format{
An \code{\link[MSnbase]{MSnSet-class}} object with 16443 features and
48 samples.
\subsection{\code{assayData}}{

A \code{matrix} of counts with 16443 rows and 48 columns.
}

\subsection{\code{featureData}}{

A \code{data.frame} with 30304 rows and 7 variables containing additional
feature information:
\describe{
\item{feature_ID}{character; }
\item{gene_symbol}{character list; the gene symbol associated with each
protein.}
\item{entrez_gene}{numeric list; the Entrez gene identifier associated
with each protein.}
\item{site}{character; }
\item{human_feature_ID}{character; }
\item{human_uniprot}{character; }
\item{human_site}{character; }
}

Prior to collapsing the \code{entrez_id} and \code{gene_symbol} columns into lists,
any rows where \code{gene_symbol} began with "LOC" were removed unless that
would remove entire features. This reduces the impact of the one-to-many
feature-to-gene mapping on FGSEA results.
}

\subsection{\code{phenoData}}{

A \code{data.frame} with 48 rows and 9 variables containing sample information:
\describe{
\item{viallabel}{character; unique sample identifier.}
\item{rin}{numeric; RNA integrity number (RIN).}
\item{pct_globin}{numeric; percent of reads mapping to globin.}
\item{pct_umi_dup}{numeric; percent of PCR duplicates as quantified with
Unique Molecular Identifiers (UMIs).}
\item{median_5_3_bias}{numeric; median 5'-3' bias.}
\item{bid}{character; Unique 5 digit numeric identifier of all samples
collected for an acute test/sample collection
period. All samples collected during that period will have the same BID.
Same as first 5 digits of \code{viallabel}.}
\item{sex}{factor; the sex of the rat with 2 levels "Female" and "Male".}
\item{timepoint}{factor; exercise training group. Either "SED"
(sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").}
\item{exp_group}{factor; unique combination of \code{sex} (first letter) and
\code{timepoint}.}
}
}
}
\usage{
TRNSCRPT_MSNSET
}
\description{
An \code{\link[MSnbase]{MSnSet-class}} object containing
transcriptomics data from rat WAT provided by MoTrPAC.
}
\details{
Two outlier samples ("90423", "90410") were removed (48 remain). Then,
\code{pct_globin}, \code{rin}, \code{pct_umi_dup}, and \code{median_5_3_bias} were mean-imputed,
centered, and scaled.
}
\examples{
TRNSCRPT_MSNSET # summary
library(MSnbase)
exprs(TRNSCRPT_MSNSET)[1:5, 1:5] # assayData
head(fData(TRNSCRPT_MSNSET))     # featureData
head(pData(TRNSCRPT_MSNSET))     # phenoData

}
\keyword{datasets}
